+++ distribution-system/src/main/java/huang/yong/chang/entity/Integral.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @TableName("tb_integral")
 public class Integral extends BaseEntity<Integral> {
 
-    public Integral(Long userId, Integer integral, Date modifyDate) {
+    public Integral(Long userId, Double integral, Date modifyDate) {
         this.userId = userId;
         this.integral = integral;
         this.modifyDate = modifyDate;
@@ -25,7 +25,7 @@ public class Integral extends BaseEntity<Integral> {
 
     //剩余积分
     @TableField("integral")
-    private Integer integral;
+    private Double integral;
 
     @TableField("modify_date")
     private Date modifyDate;
